% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_corr.R
\name{fortify_corr}
\alias{fortify_corr}
\title{Fortify a correalation test data to data.frame}
\usage{
fortify_corr(x, type = c("full", "upper", "lower"), show_diag = FALSE,
  is_corr = FALSE, corr_use = "everything", corr_method = "pearson",
  corr_test = FALSE, corr_test_alt = "two.sided",
  corr_test_method = "pearson", cluster = FALSE,
  cluster_order = "hclust", cluster_method = "complete")
}
\arguments{
\item{x}{a matrix or data.frame.}

\item{type}{a string of c("full", "upper", "lower") indicating return full matrix, upper or
lower triangle.}

\item{is_corr}{a logical value indicating whether x is a correalation matrix.}

\item{corr_use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation of)
one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs".}

\item{corr_method}{a character string indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson" (default), "kendall", or
"spearman": can be abbreviated.}

\item{corr_test}{a logical value indicating whether should compute correalation test.}

\item{corr_test_alt}{indicates the alternative hypothesis and must be one of
"two.sided", "greater" or "less". You can specify just the initial letter.
"greater" corresponds to positive association, "less" to negative association.}

\item{corr_test_method}{a character string indicating which correlation coefficient
is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{cluster}{a logical value indicating whether should reorder x by clustering.}

\item{cluster_order}{Character, the ordering method for the correlation matrix.
\itemize{
   \item{\code{"AOE"} for the angular order of the eigenvectors.
     It is calculated from the order of the angles, \eqn{a_i}:
     \deqn{ a_i = tan (e_{i2} / e_{i1}), if e_{i1} > 0}
     \deqn{ a_i = tan (e_{i2} / e_{i1}) + \pi, otherwise.}
     where \eqn{e_1} and \eqn{e_2} are the largest two eigenvalues
     of matrix \code{corr}.
     See Michael Friendly (2002) for details.}
   \item{\code{"FPC"} for the first principal component order.}
   \item{\code{"hclust"} for hierarchical clustering order.}
   \item{\code{"alphabet"} for alphabetical order.}
}}

\item{cluster_method}{Character, the agglomeration method to be used when
\code{order} is \code{hclust}. This should be one of \code{"ward"},
\code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
\code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}
}
\value{
return a data.frame.
}
\description{
Rather than generate a correalation matrix to data.frame by hand,
I recommend using the this function, becuase this function is
specially designed to accommodate other functions in ggcorrr package.
}
\details{
cor matrix is computed by \code{\link[stats]{cor}}.

correalatin test is computed by \code{\link[stats]{cor.test}},
when \code{corr_test_method = "pearson"}, the return value includes
the p values, confidence low values and upper values, other methods
include only p values.

Clustering method borrowed from \code{corrplot}, the detail please see
\code{?corrplot:::corrMatOrder}.
}
\examples{
# generate a ggcorrr_df
df <- fortify_corr(mtcars, corr_test = TRUE)

}
\seealso{
Function \code{\link{cor}}, \code{\link{cor.test}}, \code{\link{hclust}}.
}
\keyword{confidence,}
\keyword{correalation,}
\keyword{p-value,}
\keyword{significance}
