% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_lower_data}
\alias{get_lower_data}
\alias{get_upper_data}
\alias{get_diag_tri}
\alias{get_diag_data}
\alias{get_sig_data}
\title{Extracting the data subsets from a ggcorrr_df}
\usage{
get_lower_data(df, x = "x", y = "y", show_diag = TRUE)

get_upper_data(df, x = "x", y = "y", show_diag = TRUE)

get_diag_tri(df, x = "x", y = "y")

get_diag_data(df, x = "x", y = "y")

get_sig_data(df, p = "p", sig_thres = 0.05)
}
\arguments{
\item{df}{a ggcorrr_df or data.frame.}

\item{x}{the name of x coordinates in df.}

\item{y}{the name of y coordinates in df.}

\item{show_diag}{should diag rows dropped? Defaults to TRUE.}

\item{p}{the name of significant varible in df.}

\item{sig_thres}{Significance level.}
}
\value{
return a modified df.
}
\description{
\code{get_lower_data}, \code{get_upper_data} and \code{get_diag_data}
extracts left-bottom, right-top, diag rows. \code{get_diag_trip}
drop the diag rows. \code{get_sig_data} extracts the rows base on
significance test values. Except for the \code{get_sig_data} function,
other functions only support ggcorrr_df, more see @seealso \code{\link[ggcorrr]{fortify_corr}}.
}
\examples{
# generate a ggcorrr_df
df <- fortify_corr(mtcars, corr_test = TRUE)

# get lower data
get_lower_data(df)

# get upper data
get_upper_data(df)

# get diag data
get_diag_data(df)

# drop diag rows
get_diag_tri(df)

}
