% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{ggcorrr}
\alias{ggcorrr}
\title{Create a new basic plot of a correlation matrix}
\usage{
ggcorrr(data, mapping, type = c("full", "lower", "upper"),
  show_diag = TRUE, panel_backgroud = NA, grid_colour = "grey50",
  grid_size = 0.25, grid_linetype = "solid",
  axis_x_position = c("auto", "bottom", "top"),
  axis_y_position = c("auto", "left", "right"), colours = NULL,
  legend_title = NULL, legend_position = c("auto", "none", "left",
  "right", "bottom", "top"), legend_breaks = NULL,
  legend_labels = NULL, ...)
}
\arguments{
\item{data}{default dataset to use for plot.}

\item{mapping}{default list of aesthetic mappings to use for plot.}

\item{type}{character, "full" (default), "upper" or "lower", display full matrix,
lower triangular or upper triangular matrix.}

\item{show_diag}{a logical value indicating whether diag should be ploted.}

\item{panel_backgroud}{background colour of plotting area.}

\item{grid_colour}{grid line colour.}

\item{grid_size}{grid line size.}

\item{grid_linetype}{grid line type.}

\item{axis_x_position}{the position of the x axis  ("auto", "bottom", "top").}

\item{axis_y_position}{the position of the y axis  ("auto", "left", "right").}

\item{colours}{vector of colours to use for n-colour gradient.}

\item{legend_title}{a character string or expression indicating a title of guide.}

\item{legend_position}{the position of legends ("none", "auto", "left", "right", "bottom", "top",
or two-element numeric vector).}

\item{legend_breaks}{a numeric vector of positions.}

\item{legend_labels}{a character vector giving labels (must be same length as breaks).}

\item{...}{pass to `fortify_corr()` function.}
}
\description{
`ggcorrr()` initializes a ggcorrr object. It can be used to
declare the input data frame for a graphic, specify the set of plot aesthetics and
handle coordinate, colours and legend. If `data` is not a ggcorrr_df, will
call `fortify_corr()` function to trans data to ggcorrr_df.
}
\examples{
df <- fortify_corr(mtcars)
ggcorrr(df, aes(x, y)) # defaults
ggcorrr(df, aes(x, y), type = "upper") # upper triangle
ggcorrr(df, aes(x, y), type = "lower") # lower triangle
}
