% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-corr.R
\name{theme_corr}
\alias{theme_corr}
\title{ggcorrr themes}
\usage{
theme_corr(legend.position = "right", ...)
}
\arguments{
\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector).}

\item{...}{other params pass to \code{\link[ggplot2]{theme}}.}
}
\description{
These are ggcorrr themes which control
all non-data display.
}
\examples{
df <- fortify_corr(mtcars)
ggplot(df, aes(x, y, r = r)) + geom_circle2() +
coord_fixed() + theme_corr()
}
